Remove inline JS check for window.Intl
window.Intl is quite good supported. If you care about browsers that don't support it, it would IMO be better to use a polyfill for it. The advantage of removing this check is streamlining the network request waterfall, all 5 external resources defined in the HTML are now requested at the same time.